chatbot_service/requirements.txt
Flask
faiss-cpu
numpy
python-dotenv
google-generativeai
